payment-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
requests==2.31.0
pydantic==2.7.1
python-dotenv==1.0.0
python-multipart==0.0.6